fix(scoring): enforce pure integer 1 and 0 scores
- Changed Reward and State models to use int instead of float for score fields.
- Updated grader and environment logic to return and sum exact integers.
- Ensures JSON output contains 1 and 0 instead of 1.0 and 0.0.

+++ app/models.py
@@ -60,8 +60,8 @@ class Action(BaseModel):
 
 class Reward(BaseModel):
     """Rich feedback signal."""
-    score: float
-    cumulative_score: float = 0.0
+    score: int
+    cumulative_score: int = 0
     feedback: str = ""
     components: Dict[str, float] = {}
 
@@ -72,6 +72,6 @@ class State(BaseModel):
     current_index: int
     total_emails: int
     history: List[Dict[str, Any]]
-    cumulative_score: float
+    cumulative_score: int
     done: bool
     metadata: Dict[str, Any] = {}
